Do not perform deep scan when assembling Dev Cordap Jar.
Note: It is still possible to list every necessary package explicitly.
This is important as package name of the integration test is automatically included into list of packages to scan.
E.g. In case of integration test `DistributedServiceTests` the package is `net.corda.node.services` which is quite extensive, resulting big attachments (multiple MB) to be produced
and handled. This creates a notable delay in test execution.